requirements.txt
huggingface-hub
python-dotenv
streamlit
Pillow
requests
